server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.78.0",
    "fastapi>=0.128.4",
    "google-api-python-client>=2.189.0",
    "google-auth-oauthlib>=1.2.4",
    "pydantic-ai>=1.56.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
